compiler: remove *_{at_most,newer_than}_version()
The compiler version macros append an implicit .0 if the third argument
is missing, and an implicit .0.0 if the second is, too. The predictable
consequence was that I played myself assuming clang_at_most_version(13)
is equivalent to __clang_major__ <= 13, when in fact it only matches up
to exactly 13.0.0, but not e.g. 13.0.1 or newer.

To uphold the tradition of being too clever for my own good, instead of
removing the implicit version completion feature, I'll delete the worst
foot-gun macros, only keeping ones whose behavior resembles the begin()
and end() of C++ iterators.